+++ src/FlatParse/Examples/BasicLambda/Lexer.hs
@@ -11,6 +11,7 @@ module FlatParse.Examples.BasicLambda.Lexer where
 import FlatParse.Basic hiding ( Parser, runParser )
 
 import qualified FlatParse.Basic as FP
+import qualified FlatParse.BasicString as FP
 import qualified Data.ByteString as B
 import Language.Haskell.TH
 
@@ -65,7 +66,7 @@ prettyError b e =
       pos      = case e of Imprecise pos e -> pos
                            Precise pos e   -> pos
       ls       = FP.lines b
-      (l, c)   = head $ posLineCols b [pos]
+      (l, c)   = head $ FP.posLineCols b [pos]
       line     = if l < length ls then ls !! l else ""
       linum    = show l
       lpad     = map (const ' ') linum
@@ -108,7 +109,7 @@ runParser = FP.runParser
 
 -- | Run parser, print pretty error on failure.
 testParser :: Show a => Parser a -> String -> IO ()
-testParser p str = case packUTF8 str of
+testParser p str = case FP.packUTF8 str of
   b -> case runParser p b of
     Err e  -> putStrLn $ prettyError b e
     OK a _ -> print a
